requirements.txt
python-can
numpy
pandas
scikit-learn
matplotlib
seaborn
cryptography
pyyaml
joblib
pytest
colorlog
google-generativeai